+++ Shared/IPC/AppExecutionRequest.swift
@@ -67,6 +67,7 @@ public struct AppExecutionRequest: Codable, Sendable {
     public var destinationPath: String?
     public var iconSystemImageName: String?
     public var iconColorName: String?
+    public var iconImageData: Data?
     public var toolboxOption: String?
     public var openAfterCreate: Bool
     public var playSoundAfterCreate: Bool
@@ -85,6 +86,7 @@ public struct AppExecutionRequest: Codable, Sendable {
         case destinationPath
         case iconSystemImageName
         case iconColorName
+        case iconImageData
         case toolboxOption
         case openAfterCreate
         case playSoundAfterCreate
@@ -104,6 +106,7 @@ public struct AppExecutionRequest: Codable, Sendable {
         destinationPath: String? = nil,
         iconSystemImageName: String? = nil,
         iconColorName: String? = nil,
+        iconImageData: Data? = nil,
         toolboxOption: String? = nil,
         openAfterCreate: Bool = false,
         playSoundAfterCreate: Bool = false
@@ -121,6 +124,7 @@ public struct AppExecutionRequest: Codable, Sendable {
         self.destinationPath = destinationPath
         self.iconSystemImageName = iconSystemImageName
         self.iconColorName = iconColorName
+        self.iconImageData = iconImageData
         self.toolboxOption = toolboxOption
         self.openAfterCreate = openAfterCreate
         self.playSoundAfterCreate = playSoundAfterCreate
@@ -141,6 +145,7 @@ public struct AppExecutionRequest: Codable, Sendable {
         self.destinationPath = try container.decodeIfPresent(String.self, forKey: .destinationPath)
         self.iconSystemImageName = try container.decodeIfPresent(String.self, forKey: .iconSystemImageName)
         self.iconColorName = try container.decodeIfPresent(String.self, forKey: .iconColorName)
+        self.iconImageData = try container.decodeIfPresent(Data.self, forKey: .iconImageData)
         self.toolboxOption = try container.decodeIfPresent(String.self, forKey: .toolboxOption)
         self.openAfterCreate = try container.decodeIfPresent(Bool.self, forKey: .openAfterCreate) ?? false
         self.playSoundAfterCreate = try container.decodeIfPresent(Bool.self, forKey: .playSoundAfterCreate) ?? false